Add generic CloudStore methods
We allow dispatching on either the store type (AWS.Bucket or Azure.Container for now)
or you can just provide a cloud url and we'll try to figure out where to direct it.

I think this API has unfolded really nicely because you can imagine this being really convenient
for like, kaggle competitions where I have a public s3 url that I'd like to be able to use
the concurrent downloading capabilities with and just call `CloudStore.get(url)`.

This also provides a couple of options for workflows where you really want to interface w/ different
cloud providers: provide the different urls or construct the cloud-specific stores and then dispatch.